requirements.txt
aiofiles==22.1.0
aiohttp==3.9.0
aiosignal==1.2.0
aiosqlite==0.18.0
anyio==3.5.0
appdirs==1.4.4
archspec==0.2.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==4.1.0
boltons==23.0.0
Bottleneck==1.3.7
Brotli==1.0.9
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
comm==0.1.2
conda==23.11.0
conda-libmamba-solver==23.12.0
conda-package-handling==2.2.0
conda_package_streaming==0.9.0
contourpy==1.2.0
cramjam==2.7.0
cryptography==41.0.7
cycler==0.11.0
datasets==2.12.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
detoxify==0.5.2
dill==0.3.6
distro==1.8.0
entrypoints==0.4
exceptiongroup==1.0.4
executing==0.8.3
fastjsonschema==2.16.2
fastparquet==0.0.0
filelock==3.13.1
fonttools==4.25.0
frozenlist==1.4.0
fsspec==2023.10.0
gensim==4.3.2
Glances==3.4.0.3
gmpy2==2.1.2
h11==0.14.0
httpcore==1.0.4
httpx==0.25.2
huggingface-hub==0.20.3
idna==3.4
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.25.0
ipython==8.20.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.6
jsonpatch==1.32
jsonpointer==2.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter_client==8.6.0
jupyter-console==6.6.3
jupyter_core==5.5.0
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_server==2.10.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyter-ydoc==0.2.4
jupyterlab==4.0.8
jupyterlab-pygments==0.1.2
jupyterlab_server==2.25.1
jupyterlab-widgets==3.0.9
kiwisolver==1.4.4
libmambapy==1.5.6
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
menuinst==2.0.1
mistune==2.0.4
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
munkres==1.1.4
mypy==1.5.1
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numexpr==2.8.7
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
ollama==0.1.7
OpenAttack==2.1.1
overrides==7.4.0
packaging==23.1
pandas==2.2.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pip==23.3.1
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
pluggy==1.0.0
ply==3.11
pooch==1.4.0
prometheus-client==0.14.1
prompt-toolkit==3.0.43
protobuf==5.26.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py3nvml==0.2.7
pyarrow==11.0.0
pycosat==0.6.6
pycparser==2.21
Pygments==2.15.1
pynvml==11.5.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.0
QtPy==2.4.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
responses==0.13.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
safetensors==0.4.2
scikit-learn==1.3.0
scipy==1.11.4
seaborn==0.12.2
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools==68.2.2
sip==6.7.12
six==1.16.0
smart-open==7.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.2.0
sympy==1.12
terminado==0.17.1
threadpoolctl==2.2.0
tinycss2==1.2.1
tokenizers==0.15.1
tomli==2.0.1
torch==2.2.1+cu118
torchaudio==2.1.1
torchvision==0.16.1
tornado==6.3.3
tqdm==4.65.0
traitlets==5.7.1
transformers==4.38.2
triton==2.2.0
truststore==0.8.0
typing_extensions==4.9.0
tzdata==2023.3
ujson==5.9.0
urllib3==1.26.18
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
wheel==0.41.2
widgetsnbextension==4.0.5
wrapt==1.16.0
xmltodict==0.13.0
xxhash==2.0.2
y-py==0.5.9
yarl==1.9.3
ypy-websocket==0.8.2
zipp==3.17.0
zstandard==0.19.0
